+++ src/FluentlyHttpClient/FluentHttpClient.cs
@@ -7,20 +7,71 @@ using System.Net.Http;
 using System.Net.Http.Formatting;
 using System.Net.Http.Headers;
 using System.Threading.Tasks;
+using FluentlyHttpClient.Middleware;
 
 namespace FluentlyHttpClient
 {
-	public class FluentHttpClientOptions
+	public interface IFluentHttpClient
 	{
-		public string BaseUrl { get; set; }
-		public TimeSpan Timeout { get; set; }
-		public string Identifier { get; set; }
-		public Dictionary<string, string> Headers { get; set; }
-		public List<Type> Middleware { get; set; }
+		/// <summary>
+		/// Get the identifier (key) for this instance, which is registered with, within the factory.
+		/// </summary>
+		string Identifier { get; }
+
+		/// <summary>
+		/// Gets the base uri address for each request.
+		/// </summary>
+		string BaseUrl { get; }
+
+		/// <summary>
+		/// Raw http client. This should be avoided from being used.
+		/// However if something is not exposed and its really needed, it can be used from here.
+		/// </summary>
+		HttpClient RawHttpClient { get; }
+
+		/// <summary>
+		/// Formatters to be used for content negotiation for "Accept" and also sending formats. e.g. (JSON, XML)
+		/// </summary>
+		MediaTypeFormatterCollection Formatters { get; }
+
+		/// <summary>
+		/// Gets the headers which should be sent with each request.
+		/// </summary>
+		HttpRequestHeaders Headers { get; }
+
+		/// <summary>Get the formatter for an HTTP content type.</summary>
+		/// <param name="contentType">The HTTP content type (or <c>null</c> to automatically select one).</param>
+		/// <exception cref="System.InvalidOperationException">No MediaTypeFormatters are available on the API client for this content type.</exception>
+		MediaTypeFormatter GetFormatter(MediaTypeHeaderValue contentType = null);
+
+		/// <summary>
+		/// Create a new request builder which can be configured fluently.
+		/// </summary>
+		/// <param name="uriTemplate">Uri resource template e.g. <c>"/org/{id}"</c></param>
+		/// <param name="interpolationData">Data to interpolate within the Uri template place holders e.g. <c>{id}</c>. Can be either dictionary or object.</param>
+		/// <returns>Returns a new request builder.</returns>
+		FluentHttpRequestBuilder CreateRequest(string uriTemplate = null, object interpolationData = null);
+
+		/// <summary>
+		/// Build and send HTTP request.
+		/// </summary>
+		/// <param name="builder">Request builder to build request from.</param>
+		/// <returns>Returns http response.</returns>
+		Task<FluentHttpResponse> Send(FluentHttpRequestBuilder builder);
+
+		/// <summary>
+		/// Send HTTP request.
+		/// </summary>
+		/// <param name="fluentRequest">HTTP fluent request to send.</param>
+		/// <returns>Returns http response.</returns>
+		Task<FluentHttpResponse> Send(FluentHttpRequest fluentRequest);
 	}
 
+	/// <summary>
+	/// Provides a class for sending HTTP requests with a high level fluent API.
+	/// </summary>
 	[DebuggerDisplay("{DebuggerDisplay,nq}")]
-	public class FluentHttpClient
+	public class FluentHttpClient : IFluentHttpClient
 	{
 		private string DebuggerDisplay => $"[{Identifier}] BaseUrl: '{BaseUrl}', MiddlewareCount: {_middleware.Count}";
 
@@ -29,6 +80,9 @@ namespace FluentlyHttpClient
 		/// </summary>
 		public string Identifier { get; }
 
+		/// <summary>
+		/// Gets the base uri address for each request.
+		/// </summary>
 		public string BaseUrl { get; }
 
 		/// <summary>
@@ -37,15 +91,22 @@ namespace FluentlyHttpClient
 		/// </summary>
 		public HttpClient RawHttpClient { get; }
 
+		/// <summary>
+		/// Formatters to be used for content negotiation for "Accept" and also sending formats. e.g. (JSON, XML)
+		/// </summary>
 		public MediaTypeFormatterCollection Formatters { get; } = new MediaTypeFormatterCollection();
+
+		/// <summary>
+		/// Gets the headers which should be sent with each request.
+		/// </summary>
 		public HttpRequestHeaders Headers { get; }
 
+		private readonly Action<FluentHttpRequestBuilder> _requestBuilderDefaults;
 		private readonly IServiceProvider _serviceProvider;
 		private readonly IFluentHttpMiddlewareRunner _middlewareRunner;
 		private readonly IList<Type> _middleware;
 
-		public FluentHttpClient(FluentHttpClientOptions options, IServiceProvider serviceProvider,
-			IFluentHttpMiddlewareRunner middlewareRunner)
+		public FluentHttpClient(FluentHttpClientOptions options, IServiceProvider serviceProvider, IFluentHttpMiddlewareRunner middlewareRunner)
 		{
 			_serviceProvider = serviceProvider;
 			_middlewareRunner = middlewareRunner;
@@ -54,29 +115,7 @@ namespace FluentlyHttpClient
 			_middleware = options.Middleware;
 			Identifier = options.Identifier;
 			BaseUrl = options.BaseUrl;
-		}
-
-		public Task<T> Get<T>(string url)
-		{
-			return CreateRequest(url)
-				.AsGet()
-				.Return<T>();
-		}
-
-		public Task<T> Post<T>(string url, object data, MediaTypeHeaderValue contentType = null)
-		{
-			return CreateRequest(url)
-				.AsPost()
-				.WithBody(data, contentType)
-				.Return<T>();
-		}
-
-		public Task<T> Patch<T>(string url, object data, MediaTypeHeaderValue contentType = null)
-		{
-			return CreateRequest(url)
-				.AsPatch()
-				.WithBody(data, contentType)
-				.Return<T>();
+			_requestBuilderDefaults = options.RequestBuilderDefaults;
 		}
 
 		/// <summary>Get the formatter for an HTTP content type.</summary>
@@ -96,20 +135,54 @@ namespace FluentlyHttpClient
 			return formatter;
 		}
 
+		/// <summary>
+		/// Create a new request builder which can be configured fluently.
+		/// </summary>
+		/// <param name="uriTemplate">Uri resource template e.g. <c>"/org/{id}"</c></param>
+		/// <param name="interpolationData">Data to interpolate within the Uri template place holders e.g. <c>{id}</c>. Can be either dictionary or object.</param>
+		/// <returns>Returns a new request builder.</returns>
 		public FluentHttpRequestBuilder CreateRequest(string uriTemplate = null, object interpolationData = null)
 		{
 			var builder = ActivatorUtilities.CreateInstance<FluentHttpRequestBuilder>(_serviceProvider, this);
+			_requestBuilderDefaults?.Invoke(builder);
 			return uriTemplate != null
 				? builder.WithUri(uriTemplate, interpolationData)
 				: builder;
 		}
 
+		/// <summary>
+		/// Build and send HTTP request.
+		/// </summary>
+		/// <param name="builder">Request builder to build request from.</param>
+		/// <returns>Returns http response.</returns>
+		public Task<FluentHttpResponse> Send(FluentHttpRequestBuilder builder) => Send(builder.Build());
+
+		/// <summary>
+		/// Send HTTP request.
+		/// </summary>
+		/// <param name="fluentRequest">HTTP fluent request to send.</param>
+		/// <returns>Returns http response.</returns>
+		public async Task<FluentHttpResponse> Send(FluentHttpRequest fluentRequest)
+		{
+			if (fluentRequest == null) throw new ArgumentNullException(nameof(fluentRequest));
+			var response = await _middlewareRunner.Run(_middleware, fluentRequest, async request =>
+			{
+				var result = await RawHttpClient.SendAsync(request.RawRequest, request.CancellationToken).ConfigureAwait(false);
+				return ToFluentResponse(result);
+			}).ConfigureAwait(false);
+
+			if (fluentRequest.HasSuccessStatusOrThrow)
+				response.EnsureSuccessStatusCode();
+
+			return response;
+		}
+
 		private HttpClient Configure(FluentHttpClientOptions options)
 		{
-			var httpClient = new HttpClient
-			{
-				BaseAddress = new Uri(options.BaseUrl)
-			};
+			var httpClient = options.HttpMessageHandler == null
+				? new HttpClient()
+				: new HttpClient(options.HttpMessageHandler);
+			httpClient.BaseAddress = new Uri(options.BaseUrl);
 			httpClient.DefaultRequestHeaders.Add("Accept", Formatters.SelectMany(x => x.SupportedMediaTypes).Select(x => x.MediaType));
 			httpClient.Timeout = options.Timeout;
 
@@ -119,25 +192,8 @@ namespace FluentlyHttpClient
 			return httpClient;
 		}
 
-		public Task<FluentHttpResponse<T>> Send<T>(FluentHttpRequestBuilder builder) => Send<T>(builder.Build());
-
-		public async Task<FluentHttpResponse<T>> Send<T>(FluentHttpRequest fluentRequest)
-		{
-			if (fluentRequest == null) throw new ArgumentNullException("fluentRequest");
-			var response = await _middlewareRunner.Run<T>(_middleware, fluentRequest, async request =>
-			{
-				var result = await RawHttpClient.SendAsync(request.RawRequest);
-				return ToFluentResponse<T>(result);
-			});
-
-			// todo: implement this better
-			response.EnsureSuccessStatusCode();
-
-			return (FluentHttpResponse<T>)response;
-		}
-
-		private static FluentHttpResponse<T> ToFluentResponse<T>(HttpResponseMessage response) =>
-			new FluentHttpResponse<T>(response);
+		private static FluentHttpResponse ToFluentResponse(HttpResponseMessage response) =>
+			new FluentHttpResponse(response);
 
 	}
 }