+++ src/FluentlyHttpClient/FluentHttpRequestBuilder.cs
@@ -4,16 +4,20 @@ using System.Net.Http;
 using System.Net.Http.Formatting;
 using System.Net.Http.Headers;
 using System.Text.RegularExpressions;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace FluentlyHttpClient
 {
+	/// <summary>
+	/// Class to build <see cref="FluentHttpRequest"/> with a fluent API.
+	/// </summary>
 	public class FluentHttpRequestBuilder
 	{
 		/// <summary>
 		/// Gets the Http Method for the Http Request.
 		/// </summary>
-		public HttpMethod HttpMethod { get; private set; }
+		public HttpMethod HttpMethod { get; private set; } = HttpMethod.Get;
 
 		/// <summary>
 		/// Gets the Uri used for the HTTP request.
@@ -25,15 +29,20 @@ namespace FluentlyHttpClient
 		/// </summary>
 		public string UriTemplate { get; private set; }
 
+		/// <summary>
+		/// Get the headers to be sent with this request.
+		/// </summary>
 		public Dictionary<string, string> Headers { get; private set; }
 
-		private readonly FluentHttpClient _fluentHttpClient;
+		private readonly IFluentHttpClient _fluentHttpClient;
 		private HttpContent _httpBody;
 		private static readonly Regex InterpolationRegex = new Regex(@"\{(\w+)\}", RegexOptions.Compiled);
 		private object _queryParams;
 		private bool _lowerCaseQueryKeys;
+		private bool _hasSuccessStatusOrThrow;
+		private CancellationToken _cancellationToken;
 
-		public FluentHttpRequestBuilder(FluentHttpClient fluentHttpClient)
+		public FluentHttpRequestBuilder(IFluentHttpClient fluentHttpClient)
 		{
 			_fluentHttpClient = fluentHttpClient;
 		}
@@ -66,7 +75,7 @@ namespace FluentlyHttpClient
 		}
 
 		/// <summary>
-		/// Add the specified header and its value for the request.
+		/// Add the specified headers and their value for the request.
 		/// </summary>
 		/// <param name="headers">Headers to add.</param>
 		/// <returns>Returns request builder for chaining.</returns>
@@ -107,7 +116,7 @@ namespace FluentlyHttpClient
 
 		/// <summary>Set the body content of the HTTP request.</summary>
 		/// <param name="body">Value to serialize into the HTTP body content.</param>
-		/// <param name="contentType">Request body format (or <c>null</c> to use the first supported Content-Type in the <see cref="FluentHttpClient.Formatters"/>).</param>
+		/// <param name="contentType">Request body format (or <c>null</c> to use the first supported Content-Type in the <see cref="IFluentHttpClient.Formatters"/>).</param>
 		/// <returns>Returns the request builder for chaining.</returns>
 		/// <exception cref="InvalidOperationException">No MediaTypeFormatters are available on the API client for this content type.</exception>
 		public FluentHttpRequestBuilder WithBody<T>(T body, MediaTypeHeaderValue contentType = null)
@@ -119,7 +128,7 @@ namespace FluentlyHttpClient
 
 		/// <summary>Set the body content of the HTTP request.</summary>
 		/// <param name="body">Value to serialize into the HTTP body content.</param>
-		/// <param name="contentType">Request body format (or <c>null</c> to use the first supported Content-Type in the <see cref="FluentHttpClient.Formatters"/>).</param>
+		/// <param name="contentType">Request body format (or <c>null</c> to use the first supported Content-Type in the <see cref="IFluentHttpClient.Formatters"/>).</param>
 		/// <exception cref="InvalidOperationException">No MediaTypeFormatters are available on the API client for this content type.</exception>
 		/// <returns>Returns the request builder for chaining.</returns>
 		public FluentHttpRequestBuilder WithBody(object body, MediaTypeHeaderValue contentType = null)
@@ -158,19 +167,60 @@ namespace FluentlyHttpClient
 			return this;
 		}
 
+		/// <summary>Determine whether the status code should succeeds or else throw.</summary>
+		/// <param name="hasSuccessStatusOrThrow">When true status should succeed otherwise it will throw.</param>
+		/// <returns>Returns the request builder for chaining.</returns>
+		public FluentHttpRequestBuilder WithSuccessStatus(bool hasSuccessStatusOrThrow = true)
+		{
+			_hasSuccessStatusOrThrow = hasSuccessStatusOrThrow;
+			return this;
+		}
+
+		/// <summary>.</summary>
+		/// <returns>Returns the request builder for chaining.</returns>
+		public FluentHttpRequestBuilder WithCancellationToken(CancellationToken cancellationToken)
+		{
+			_cancellationToken = cancellationToken;
+			return this;
+		}
+		
+		/// <summary>
+		/// Send request, read content with the type specified (when success) and return data directly.
+		/// </summary>
+		/// <typeparam name="T">Type to return.</typeparam>
+		/// <returns></returns>
 		public async Task<T> Return<T>()
 		{
 			var response = await ReturnAsResponse<T>();
 			return response.Data;
 		}
 
+		/// <summary>
+		/// Send request and returns Http Response and also read content with the type specified (when success).
+		/// </summary>
+		/// <typeparam name="T">Type to return.</typeparam>
+		/// <returns>Return response with data typed.</returns>
 		public async Task<FluentHttpResponse<T>> ReturnAsResponse<T>()
 		{
-			var response = await _fluentHttpClient.Send<T>(this);
-			response.Data = await response.RawResponse.Content.ReadAsAsync<T>(_fluentHttpClient.Formatters);
-			return response;
+			var response = await ReturnAsResponse();
+			var genericResponse = new FluentHttpResponse<T>(response);
+
+			if (genericResponse.IsSuccessStatusCode)
+				genericResponse.Data = await genericResponse.RawResponse.Content.ReadAsAsync<T>(_fluentHttpClient.Formatters, _cancellationToken);
+			
+			return genericResponse;
 		}
 
+		/// <summary>
+		/// Send request and returns Http Response.
+		/// </summary>
+		/// <returns>Returns an http response.</returns>
+		public async Task<FluentHttpResponse> ReturnAsResponse() => await _fluentHttpClient.Send(this);
+
+		/// <summary>
+		/// Build http request.
+		/// </summary>
+		/// <returns>Return http request instance.</returns>
 		public FluentHttpRequest Build()
 		{
 			ValidateRequest();
@@ -184,10 +234,18 @@ namespace FluentlyHttpClient
 				foreach (var header in Headers)
 					httpRequest.Headers.Add(header.Key, header.Value);
 
-			var fluentRequest = new FluentHttpRequest(httpRequest);
+			var fluentRequest = new FluentHttpRequest(httpRequest)
+			{
+				HasSuccessStatusOrThrow = _hasSuccessStatusOrThrow,
+				CancellationToken = _cancellationToken
+			};
 			return fluentRequest;
 		}
 
+		/// <summary>
+		/// Ensure validate request or throw.
+		/// </summary>
+		/// <exception cref="RequestValidationException">When request is not valid.</exception>
 		protected void ValidateRequest()
 		{
 			if (HttpMethod == null)