+++ src/FluentlyHttpClient/IFluentHttpMiddleware.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace FluentlyHttpClient
-{
-	public interface IFluentHttpMiddleware
-	{
-		Task<IFluentHttpResponse> Invoke(FluentHttpRequest request);
-	}
-}