+++ src/FluentlyHttpClient/Middleware/IFluentHttpMiddleware.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace FluentlyHttpClient.Middleware
+{
+	public interface IFluentHttpMiddleware
+	{
+		Task<FluentHttpResponse> Invoke(FluentHttpRequest request);
+	}
+}