+++ src/FluentlyHttpClient/Middleware/TimerHttpMiddleware.cs
@@ -1,33 +1,14 @@
-﻿using Microsoft.Extensions.Logging;
+﻿using FluentlyHttpClient.Middleware;
+using Microsoft.Extensions.Logging;
 using System;
 using System.Diagnostics;
 using System.Threading.Tasks;
 
-namespace FluentlyHttpClient
+namespace FluentlyHttpClient.Middleware
 {
-	public class LoggerHttpMiddleware : IFluentHttpMiddleware
-	{
-		private readonly FluentHttpRequestDelegate _next;
-		private readonly ILogger _logger;
-
-		public LoggerHttpMiddleware(FluentHttpRequestDelegate next, ILogger<LoggerHttpMiddleware> logger)
-		{
-			_next = next;
-			_logger = logger;
-		}
-
-		public async Task<IFluentHttpResponse> Invoke(FluentHttpRequest request)
-		{
-			if (_logger.IsEnabled(LogLevel.Information))
-				_logger.LogInformation("Pre-request... {request}", request);
-
-			var response = await _next(request);
-
-			if (_logger.IsEnabled(LogLevel.Information))
-				_logger.LogInformation("Post-request... {response}", response);
-			return response;
-		}
-	}
+	/// <summary>
+	/// Timer middleware for Http client.
+	/// </summary>
 	public class TimerHttpMiddleware : IFluentHttpMiddleware
 	{
 		private readonly FluentHttpRequestDelegate _next;
@@ -39,7 +20,7 @@ namespace FluentlyHttpClient
 			_logger = logger;
 		}
 
-		public async Task<IFluentHttpResponse> Invoke(FluentHttpRequest request)
+		public async Task<FluentHttpResponse> Invoke(FluentHttpRequest request)
 		{
 			var watch = Stopwatch.StartNew();
 			var response = await _next(request);
@@ -56,17 +37,30 @@ namespace FluentlyHttpClient
 			return response;
 		}
 	}
+}
 
-	public static class FluentResponseExtensions
+namespace FluentlyHttpClient
+{
+	public static class TimerFluentResponseExtensions
 	{
 		private const string TimeTakenKey = "TIME_TAKEN";
 
-		public static void SetTimeTaken(this IFluentHttpResponse response, TimeSpan value)
+		/// <summary>
+		/// Set time taken.
+		/// </summary>
+		/// <param name="response">Response instance.</param>
+		/// <param name="value">Timespan value.</param>
+		public static void SetTimeTaken(this FluentHttpResponse response, TimeSpan value)
 		{
 			response.Items.Add(TimeTakenKey, value);
 		}
 
-		public static TimeSpan GetTimeTaken(this IFluentHttpResponse response)
+		/// <summary>
+		/// Get time taken for the response. This is generally set via <see cref="TimerHttpMiddleware"/>.
+		/// </summary>
+		/// <param name="response">Response to get time from.</param>
+		/// <returns>Returns timespan for the time taken.</returns>
+		public static TimeSpan GetTimeTaken(this FluentHttpResponse response)
 		{
 			return (TimeSpan)response.Items[TimeTakenKey];
 		}