+++ test/FluentHttpClientFactoryTest.cs
@@ -1,0 +1,72 @@
+﻿using FluentlyHttpClient;
+using System.Net.Http;
+using Xunit;
+using static FluentlyHttpClient.Test.ServiceTestUtil;
+
+namespace Test
+{
+	public class ClientFactory_WithRequestBuilderDefaults
+	{
+		[Fact]
+		public void WithCustomRequestDefaults()
+		{
+			var fluentHttpClientFactory = GetNewClientFactory();
+			fluentHttpClientFactory.CreateBuilder("abc")
+				.WithBaseUrl("http://abc.com")
+				.WithRequestBuilderDefaults(builder => builder.AsPut())
+				.Register();
+
+			var httpClient = fluentHttpClientFactory.Get("abc");
+			var request = httpClient.CreateRequest("/api")
+				.Build();
+
+			Assert.NotNull(request);
+			Assert.Equal(HttpMethod.Put, request.Method);
+		}
+	}
+
+	public class ClientFactory_Register
+	{
+		[Fact]
+		public void ShouldRegisterSuccessfully()
+		{
+			var fluentHttpClientFactory = GetNewClientFactory();
+			fluentHttpClientFactory.CreateBuilder("abc")
+				.WithBaseUrl("http://abc.com")
+				.Register();
+
+			var httpClient = fluentHttpClientFactory.Get("abc");
+			Assert.NotNull(httpClient);
+			Assert.Equal("abc", httpClient.Identifier);
+		}
+
+		[Fact]
+		public void ThrowsErrorWhenIdentifierNotSpecified()
+		{
+			var fluentHttpClientFactory = GetNewClientFactory();
+			var clientBuilder = fluentHttpClientFactory.CreateBuilder(null);
+			Assert.Throws<ClientBuilderValidationException>(() => clientBuilder.Register());
+		}
+
+		[Fact]
+		public void ThrowsErrorWhenUriNotSpecified()
+		{
+			var fluentHttpClientFactory = GetNewClientFactory();
+			var clientBuilder = fluentHttpClientFactory.CreateBuilder("abc");
+
+			Assert.Throws<ClientBuilderValidationException>(() => clientBuilder.Register());
+		}
+
+		[Fact]
+		public void ThrowsErrorWhenAlreadyRegistered()
+		{
+			var fluentHttpClientFactory = GetNewClientFactory();
+			var clientBuilder = fluentHttpClientFactory.CreateBuilder("abc")
+					.WithBaseUrl("http://abc.com")
+					.Register();
+
+			Assert.Throws<ClientBuilderValidationException>(() => clientBuilder.Register());
+		}
+	}
+
+}